// data/enums/unit.ts
export enum Unit {
  PIECE = 'PIECE', // 개
  GRAM = 'GRAM', // g
  KILOGRAM = 'KILOGRAM', // kg
  MILLILITER = 'MILLILITER', // ml
  LITER = 'LITER', // L
  PACK = 'PACK', // 팩
  BUNCH = 'BUNCH', // 묶음
  WHOLE = 'WHOLE', // 마리
  EGG = 'EGG', // 알
  SLICE = 'SLICE', // 조각
  HEAD = 'HEAD', // 포기
  SHEET = 'SHEET', // 장
  CLUSTER = 'CLUSTER', // 송이
  CAN = 'CAN', // 캔
  PLANT = 'PLANT', // 모
}